Firebase Authentication and User Registration/Login/Password Reset
This is the name of the Firebase project I created for our group to host authentication for login. I also created the registration, login and password reset pages for our website. I have tested them multiple times and they are working for all of the features mentioned above. Added password validation on 6/9/2022.