Firebase authentication v2
Created new login/registration pages with a temporary home page to test the firebase authentication. The user is able to register and login and the webpage is able to know if the user is logged in or not. Current issues are that there needs to be a connect between Firebase authenication and the Firebase database as the current login process goes through the database (previously it was through Firebase authentication.) After solving this I will need to implement these features into the new webpage UI I have started.