Firebase Authentication and Database sync
Created a new folder for images..
Setup a newer version of the register and login pages which will register the user by UID and not username in the database. This is important because with the UID we can limit the amount of accounts that 1 email can have. I am still working on having unique usernames as they can be replicated at this time. I have done some modifications to the main webpage UI by creating extra cards and adding image based links to them. I will add the other kinds of validation to username/password/name at a later time, probably during sprint2.