Removed username field from all sections
I removed the username field as we are no longer using that as the main authentication point for the website. It was causing issues in firebase database because there could be duplicate usernames. Switched to email validation and there are no longer any duplicates. Testregisterv5 and teststyleregisterv4 are an attempt to style the page which is still a work in  progress.