Website Authentication changes
Redid Firebase authentication. The website will now track authentication (only on the main webpage at this time) for the currently logged in user. The authentication method now goes through Firebase Auth instead of a realtime Firebase database. The reason for these changes was that the change password method in Firebase goes through Firebase auth and there was a disconnect between the two passwords when trying to call this method previously. Removed the login/registration pages and replaced it with modal windows. Also added a forgot password modal window and account modal window. The account modal window is still a work. Apparently there is an issue with Email verification that is a known issue prior to registration. Because of this I may have to change authentication to be through email (so passwordless) in order to fulfill our non functional requirements. The navbar was also changed to be shown when scrolling throughout the webpage, this may be changed in the future as well.